@gogetron 2024-03-28
Hi, thank you for your review, I have applied your suggested modification as it made a lot of sense. I have followed a similar idea as in my other PR #1067. In this function I had to add the batch_size parameter.

**Code Setup**
```python
import numpy as np

from cleanlab.segmentation.summary import common_label_issues
from cleanlab.segmentation.rank import find_label_issues

SIZE = 250
NUM_IMAGES = 1000
NUM_CLASSES = 10
np.random.seed(0)
%load_ext memory_profiler

def generate_image_dataset():
    labels = np.random.randint(NUM_CLASSES, size=(NUM_IMAGES, SIZE, SIZE), dtype=int)
    pred_probs = np.random.random((NUM_IMAGES, NUM_CLASSES, SIZE, SIZE))
    return labels, pred_probs

# Create input data
labels, pred_probs = generate_image_dataset()
issues = find_label_issues(labels, pred_probs, n_jobs=1, verbose=False)
```

**Current version**

```python
%%timeit
%memit common_label_issues(issues, labels, pred_probs, verbose=False)
# peak memory: 8039.15 MiB, increment: 2437.64 MiB
# peak memory: 8039.63 MiB, increment: 2473.30 MiB
# peak memory: 8039.59 MiB, increment: 2473.26 MiB
# peak memory: 8039.46 MiB, increment: 2473.14 MiB
# peak memory: 8039.69 MiB, increment: 2473.37 MiB
# peak memory: 8039.66 MiB, increment: 2473.15 MiB
# peak memory: 8039.67 MiB, increment: 2473.16 MiB
# peak memory: 8039.69 MiB, increment: 2473.17 MiB
# 1min 11s ± 556 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
**This PR**
```python
%%timeit
%memit common_label_issues(issues, labels, pred_probs, verbose=False)
# peak memory: 7278.66 MiB, increment: 1674.31 MiB
# peak memory: 7384.47 MiB, increment: 1818.66 MiB
# peak memory: 7296.44 MiB, increment: 1730.57 MiB
# peak memory: 7296.82 MiB, increment: 1730.94 MiB
# peak memory: 7332.82 MiB, increment: 1766.75 MiB
# peak memory: 7279.39 MiB, increment: 1713.32 MiB
# peak memory: 7388.80 MiB, increment: 1822.73 MiB
# peak memory: 7332.67 MiB, increment: 1766.61 MiB
# 6.93 s ± 73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```